only for objects
